[chore] refactor for yarn workspaces
