+++ packages/client/doc/lib/mermaid-plugin.js
