+++ packages/client/doc/slides.md
