+++ packages/client/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "client",
+  "version": "1.0.0",
+  "license": "MIT",
+  "devDependencies": {
+    "@babel/core": "^7.14.3",
+    "@storybook/addon-actions": "^6.2.9",
+    "@storybook/addon-essentials": "^6.2.9",
+    "@storybook/addon-links": "^6.2.9",
+    "@storybook/react": "^6.2.9",
+    "babel-loader": "^8.2.2",
+    "http-proxy-middleware": "^2.0.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "reveal-md": "^5.1.1",
+    "storybook": "^6.2.9"
+  },
+  "dependencies": {
+    "@emotion/core": "^11.0.0",
+    "@emotion/react": "^11.4.0",
+    "@emotion/styled": "^11.3.0",
+    "@types/ws": "^7.4.5",
+    "body-parser": "^1.19.0",
+    "codemirror": "^5.61.1",
+    "node-fetch": "^2.6.1",
+    "storybook": "^6.2.9",
+    "y-codemirror": "^2.1.1",
+    "y-protocols": "^1.0.5",
+    "y-webrtc": "^10.2.0",
+    "y-websocket": "^1.3.15",
+    "yjs": "^13.5.10"
+  },
+  "peerDependencies": {
+    "react": "^17.0.2"
+  },
+  "scripts": {
+    "slides": "reveal-md doc/slides.md --scripts doc/lib/mermaid.min.js,doc/lib/mermaid-plugin.js  -w",
+    "storybook": "start-storybook -p 6006 -s ./public",
+    "prebuild": "yarn slides --static public/slides",
+    "build": "build-storybook -s ./public -o ./dist/webpage"
+  }
+}
