+++ packages/client/src/stories/editor.stories.mdx
@@ -1,6 +1,9 @@
 import { Meta } from "@storybook/addon-docs/blocks";
+import Me from "../me";
 import Editor from "../editor";
 
 <Meta title="Editor" />
+
+You are currently logged in as "<Me />"
 
 <Editor></Editor>
