+++ packages/client/src/stories/login.stories.mdx
@@ -13,5 +13,4 @@ user.
 
 <Login />
 
-
 You are currently logged in as "<Me />"
