+++ packages/client/src/stories/readme.stories.mdx
@@ -1,9 +1,8 @@
 import { Meta } from "@storybook/addon-docs/blocks";
 import { Description } from "@storybook/addon-docs/blocks";
-import Readme from "../../readme.md";
+import Readme from "../../../../readme.md";
 import Editor from "../editor";
 
 <Meta title="README" />
 
 <Description>{Readme}</Description>
-
