+++ packages/service/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "service",
+  "version": "1.0.0",
+  "license": "MIT",
+  "devDependencies": {
+    "nodemon": "^2.0.9"
+  },
+  "dependencies": {
+    "@types/ws": "^7.4.5",
+    "cookie-parser": "^1.4.5",
+    "express": "^4.17.1",
+    "express-flash": "^0.0.2",
+    "express-session": "^1.17.2",
+    "http": "^0.0.1-security",
+    "lib0": "^0.2.42",
+    "passport": "^0.4.1",
+    "passport-local": "^1.0.0",
+    "socket.io": "^4.1.2",
+    "ts-node": "^10.0.0",
+    "ws": "^7.4.6",
+    "y-codemirror": "^2.1.1",
+    "y-webrtc": "^10.2.0",
+    "y-websocket": "^1.3.15",
+    "yjs": "^13.5.10"
+  },
+  "scripts": {
+    "dev": "nodemon -e ts -w src -x yarn start",
+    "start": "ts-node src/websocket-service.ts"
+  }
+}
