+++ packages/service/src/utils/ws-shared-doc.ts
